+++ x-pack/platform/plugins/shared/streams_app/public/components/data_management/stream_detail_lifecycle/ilm_summary.tsx
@@ -28,8 +28,6 @@ import {
   formatNumber,
   useEuiTheme,
 } from '@elastic/eui';
-import { LocatorPublic } from '@kbn/share-plugin/common';
-import { IlmLocatorParams } from '@kbn/index-lifecycle-management-common-shared';
 import { useStreamsAppFetch } from '../../../hooks/use_streams_app_fetch';
 import { useKibana } from '../../../hooks/use_kibana';
 import { orderIlmPhases, parseDurationInSeconds } from './helpers';
@@ -39,11 +37,9 @@ import { useIlmPhasesColorAndDescription } from './hooks/use_ilm_phases_color_an
 export function IlmSummary({
   definition,
   lifecycle,
-  ilmLocator,
 }: {
   definition: IngestStreamGetResponse;
   lifecycle: IngestStreamLifecycleILM;
-  ilmLocator?: LocatorPublic<IlmLocatorParams>;
 }) {
   const {
     dependencies: {
@@ -104,7 +100,7 @@ export function IlmSummary({
           </EuiFlexItem>
 
           <EuiFlexItem grow={false}>
-            <IlmLink lifecycle={lifecycle} ilmLocator={ilmLocator} />
+            <IlmLink lifecycle={lifecycle} />
           </EuiFlexItem>
         </EuiFlexGroup>
       </EuiPanel>
