+++ x-pack/platform/plugins/shared/streams_app/public/components/data_management/stream_detail_lifecycle/index.tsx
@@ -16,11 +16,7 @@ import {
   isUnwiredStreamGetResponse,
   isWiredStreamGetResponse,
 } from '@kbn/streams-schema';
-import {
-  ILM_LOCATOR_ID,
-  IlmLocatorParams,
-  PolicyFromES,
-} from '@kbn/index-lifecycle-management-common-shared';
+import { PolicyFromES } from '@kbn/index-lifecycle-management-common-shared';
 import { i18n } from '@kbn/i18n';
 import { useAbortController } from '@kbn/react-hooks';
 import { useKibana } from '../../../hooks/use_kibana';
@@ -98,7 +94,6 @@ export function StreamDetailLifecycle({
     core: { http, notifications },
     dependencies: {
       start: {
-        share,
         streams: { streamsRepositoryClient },
       },
     },
@@ -121,8 +116,6 @@ export function StreamDetailLifecycle({
   } = useDataStreamStats({ definition });
 
   const { signal } = useAbortController();
-
-  const ilmLocator = share.url.locators.get<IlmLocatorParams>(ILM_LOCATOR_ID);
 
   const getIlmPolicies = () =>
     http.get<PolicyFromES[]>('/api/index_lifecycle_management/policies', {
@@ -177,7 +170,6 @@ export function StreamDetailLifecycle({
         updateLifecycle={updateLifecycle}
         getIlmPolicies={getIlmPolicies}
         updateInProgress={updateInProgress}
-        ilmLocator={ilmLocator}
       />
 
       <EuiPanel grow={false} hasShadow={false} hasBorder paddingSize="s">
@@ -190,7 +182,6 @@ export function StreamDetailLifecycle({
             <RetentionMetadata
               definition={definition}
               lifecycleActions={lifecycleActions}
-              ilmLocator={ilmLocator}
               openEditModal={(action) => setOpenEditModal(action)}
               isLoadingStats={isLoadingStats}
               stats={stats}
@@ -218,11 +209,7 @@ export function StreamDetailLifecycle({
           {isIlmLifecycle(definition.effective_lifecycle) ? (
             <EuiFlexItem grow={3}>
               <EuiPanel grow={true} hasShadow={false} hasBorder paddingSize="s">
-                <IlmSummary
-                  definition={definition}
-                  lifecycle={definition.effective_lifecycle}
-                  ilmLocator={ilmLocator}
-                />
+                <IlmSummary definition={definition} lifecycle={definition.effective_lifecycle} />
               </EuiPanel>
             </EuiFlexItem>
           ) : null}
