+++ x-pack/platform/plugins/shared/streams_app/public/components/esql_chart/controlled_esql_chart.tsx
@@ -171,6 +171,7 @@ export function ControlledEsqlChart<T extends string>({
         id="y-axis"
         ticks={3}
         position={Position.Left}
+        hide
         tickFormat={yTickFormat}
         labelFormat={yLabelFormat}
       />
