+++ x-pack/platform/plugins/shared/streams_app/public/components/stream_detail_overview/components/stream_stats_panel.tsx
@@ -17,31 +17,19 @@ import { css } from '@emotion/css';
 import { i18n } from '@kbn/i18n';
 import React, { ReactNode } from 'react';
 import { IngestStreamGetResponse, isDslLifecycle, isIlmLifecycle } from '@kbn/streams-schema';
-import { IlmLocatorParams } from '@kbn/index-lifecycle-management-common-shared';
 
-import { LocatorPublic } from '@kbn/share-plugin/public';
-import type { StreamDetailsResponse } from '@kbn/streams-plugin/server/routes/internal/streams/crud/route';
 import { IlmLink } from '../../data_management/stream_detail_lifecycle/ilm_link';
 import {
   formatBytes,
   formatIngestionRate,
 } from '../../data_management/stream_detail_lifecycle/helpers/format_bytes';
-import { DataStreamStats } from '../../data_management/stream_detail_lifecycle/hooks/use_data_stream_stats';
+import { useDataStreamStats } from '../../data_management/stream_detail_lifecycle/hooks/use_data_stream_stats';
 
 interface StreamStatsPanelProps {
-  definition?: IngestStreamGetResponse;
-  dataStreamStats?: DataStreamStats;
-  docCount?: StreamDetailsResponse;
-  ilmLocator?: LocatorPublic<IlmLocatorParams>;
+  definition: IngestStreamGetResponse;
 }
 
-const RetentionDisplay = ({
-  definition,
-  ilmLocator,
-}: {
-  definition?: IngestStreamGetResponse;
-  ilmLocator?: LocatorPublic<IlmLocatorParams>;
-}) => {
+const RetentionDisplay = ({ definition }: { definition: IngestStreamGetResponse }) => {
   if (!definition) return <>-</>;
 
   if (isDslLifecycle(definition.effective_lifecycle)) {
@@ -56,7 +44,7 @@ const RetentionDisplay = ({
   }
 
   if (isIlmLifecycle(definition.effective_lifecycle)) {
-    return <IlmLink lifecycle={definition.effective_lifecycle} ilmLocator={ilmLocator} />;
+    return <IlmLink lifecycle={definition.effective_lifecycle} />;
   }
 
   return <>-</>;
@@ -97,12 +85,8 @@ const StatItem = ({ label, value, withBorder = false }: StatItemProps) => {
   );
 };
 
-export function StreamStatsPanel({
-  definition,
-  dataStreamStats,
-  docCount,
-  ilmLocator,
-}: StreamStatsPanelProps) {
+export function StreamStatsPanel({ definition }: StreamStatsPanelProps) {
+  const dataStreamStats = useDataStreamStats({ definition }).stats;
   const retentionLabel = i18n.translate('xpack.streams.entityDetailOverview.retention', {
     defaultMessage: 'Data retention',
   });
@@ -128,7 +112,7 @@ export function StreamStatsPanel({
               {retentionLabel}
             </EuiText>
             <EuiText size="m">
-              <RetentionDisplay definition={definition} ilmLocator={ilmLocator} />
+              <RetentionDisplay definition={definition} />
             </EuiText>
           </EuiFlexGroup>
         </EuiPanel>
@@ -138,24 +122,15 @@ export function StreamStatsPanel({
           <EuiFlexGroup>
             <StatItem
               label={documentCountLabel}
-              value={docCount ? formatNumber(docCount.details.count || 0, 'decimal0') : '-'}
+              value={
+                dataStreamStats ? formatNumber(dataStreamStats.totalDocs || 0, 'decimal0') : '-'
+              }
             />
             <StatItem
-              label={
-                <>
-                  {storageSizeLabel}
-                  <EuiIconTip
-                    content={i18n.translate('xpack.streams.streamDetailOverview.sizeTip', {
-                      defaultMessage:
-                        'Estimated size based on the number of documents in the current time range and the total size of the stream.',
-                    })}
-                    position="right"
-                  />
-                </>
-              }
+              label={storageSizeLabel}
               value={
-                dataStreamStats && docCount
-                  ? formatBytes(getStorageSizeForTimeRange(dataStreamStats, docCount))
+                dataStreamStats && dataStreamStats.sizeBytes
+                  ? formatBytes(dataStreamStats.sizeBytes)
                   : '-'
               }
               withBorder
@@ -186,18 +161,4 @@ export function StreamStatsPanel({
       </EuiFlexItem>
     </EuiFlexGroup>
   );
-}
-
-function getStorageSizeForTimeRange(
-  dataStreamStats: DataStreamStats,
-  docCount: StreamDetailsResponse
-) {
-  const storageSize = dataStreamStats.sizeBytes;
-  const totalCount = dataStreamStats.totalDocs;
-  const countForTimeRange = docCount.details.count;
-  if (!storageSize || !totalCount || !countForTimeRange) {
-    return 0;
-  }
-  const bytesPerDoc = totalCount ? storageSize / totalCount : 0;
-  return bytesPerDoc * countForTimeRange;
 }
